backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.10
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
ruff==0.12.12
email-validator==2.2.0
